EnemyMove done (- animations)
Finished up, the enemy will flip when it encounters a ledge, and the whole sprite will flip in general when it turns (so that I don't have to set that up later when I get the animations)

The last thing I need is the animations for the enemies. Send me some animations and I'll do that when I get them.