Main Menu done*
*: Not the artwork, I just make the framework, you know the drill

Very basic skeleton, 5 buttons (4 for the three levels and tutorial, 1 for quit)

Updated the build settings, because I don't think they got saved last time

Also included something to work on in the future, with a disabling levels you haven't unlocked script in the works. It's not important right now, so I'll work on it once I get everything else done.