Win and Lose Scene's skeletons done
The skeletons for both the win and lose scenes are done (they were basically copy and pastes of each other, I just changed the text). 2 buttons: one to go back to the main menu, one to quit the game. Both of them should be operational.

Again, the rest is up to the artists. All of the menus are done, aside from that level-gray-out thing I mentioned in the last push. I dunno if I'll actually get to that, tho.