Enemy Movement done (again)
This time, I added a mechanic where the enemies get scared of the lantern's light and run away. It took a little bit of configuring to get done, but now it works, or it should, anyways.

It's not totally done; I mentioned on the Trello board that I wanted to input a final mechanic where if the enemy is scared of the light and running away, but they have nowhere to go (being cornered), they will turn and fight, like a real cornered animal. They might even get a bonus to their attack speed in this state.