Got comfortably far on the Toxic Rat
It is, arguably, the simplest of the bosses, but it's nice knowing that I got this far.

All I think I need to do is configure the canvas that displays the health bar and make the rat shoot bullets.

Last boss is the Resident and it's a kicking boss, so I'll have to find some way to make it interesting.

By the way, I'm going to take a copy of this game after we're done and work on it as a personal project, because I feel like I ran out of time and didn't really do the game justice.