Fixing a bug I made
This is why it's nice to have bug fixes. Goose found this one, where the ghost turns into the rat, because I was trying to fix the Rat AC and it turn out that I was actually fixing the Ghost AC. It should be fixed now.

Also I started work on level 3 boss but that's beside the point.